+++ src/lwip_driver.h
@@ -60,9 +60,6 @@ bool enet_has_hardware();
 // It is suggested to initialize the random number generator with srand() before
 // calling this.
 bool enet_init(const uint8_t mac[ETH_HWADDR_LEN],
-               const ip4_addr_t *ipaddr,
-               const ip4_addr_t *netmask,
-               const ip4_addr_t *gw,
                netif_ext_callback_fn callback);
 
 // Disables Ethernet.